Moves classes
