+++ server/src/main/java/com/github/jazzschmidt/apubsub/events/ClientDisconnectedEvent.java
@@ -1,4 +1,4 @@
-package com.github.jazzschmidt.apubsub;
+package com.github.jazzschmidt.apubsub.events;
 
 public class ClientDisconnectedEvent extends StompClientEvent {
     public ClientDisconnectedEvent(Object source, String sessionId) {
