+++ server/src/main/java/com/github/jazzschmidt/apubsub/events/StompClientEvent.java
@@ -1,4 +1,4 @@
-package com.github.jazzschmidt.apubsub;
+package com.github.jazzschmidt.apubsub.events;
 
 import org.springframework.context.ApplicationEvent;
