+++ server/src/test/groovy/com/github/jazzschmidt/apubsub/interceptor/UnregisteredClientsGuardSpec.groovy
@@ -1,13 +1,15 @@
-package com.github.jazzschmidt.apubsub
+package com.github.jazzschmidt.apubsub.interceptor
 
+import com.github.jazzschmidt.apubsub.ClientRegistrationService
 import com.github.jazzschmidt.apubsub.extra.MessageBuilder
+import com.github.jazzschmidt.apubsub.interceptor.UnregisteredClientsGuard
 import org.springframework.messaging.Message
 import org.springframework.messaging.MessageChannel
 import org.springframework.messaging.MessagingException
 import org.springframework.messaging.support.ChannelInterceptor
 import spock.lang.Specification
 
-class RegistrationChannelInterceptorSpec extends Specification {
+class UnregisteredClientsGuardSpec extends Specification {
 
     def 'rejects subscriptions from unregistered client'() {
         given:
@@ -15,7 +17,7 @@ class RegistrationChannelInterceptorSpec extends Specification {
         Message<?> subscribe = MessageBuilder.subscribe(sessionId)
 
         ClientRegistrationService registrationService = Mock()
-        ChannelInterceptor interceptor = new RegistrationChannelInterceptor(registrationService)
+        ChannelInterceptor interceptor = new UnregisteredClientsGuard(registrationService)
 
         when:
         interceptor.preSend(subscribe, Stub(MessageChannel))
@@ -31,7 +33,7 @@ class RegistrationChannelInterceptorSpec extends Specification {
         Message<?> subscribe = MessageBuilder.subscribe(sessionId)
 
         ClientRegistrationService registrationService = Mock()
-        ChannelInterceptor interceptor = new RegistrationChannelInterceptor(registrationService)
+        ChannelInterceptor interceptor = new UnregisteredClientsGuard(registrationService)
 
         when:
         interceptor.preSend(subscribe, Stub(MessageChannel))
